+++ contrib/old_script_to_run_plom_in_docker.sh
@@ -1,7 +1,15 @@
 #!/bin/sh
 
+# SPDX-License-Identifier: AGPL-3.0-or-later
+# Copyright (C) 2019-2021 Colin B. Macdonald
+# Copyright (C) 2020 Andrew Rechnitzer
+
 # Instructions
 # ------------
+#
+# This script runs a Plom demo in a Docker container.  Not sure its useful
+# compared to `docker run -it --rm -p ... -v ... plomgrading/server plom-demo`
+# but it remains here for reference.
 #
 # Copy this script somewhere
 #   `cp plomdock.sh ${HOME}/plomdock/`
@@ -15,14 +23,12 @@
 #   `docker build . --tag plomship`
 # in the plom source directory, where the `Dockerfile` is.
 
-# semi-regularly updated devel version:
-export BASEIMG=docker.io/cbm755/plom:devel
+# Official docker image
+export BASEIMG=docker.io/plomgrading/server
 # Local image:
 #export BASEIMG=plomship
 # Current main branch:
 #export BASEIMG=registry.gitlab.com/plom/plom:master
-# Official docker image, TODO: coming soon:
-#export BASEIMG=docker.io/plomgrading/plom
 
 export PD=plom0
 export UID=`id -u`