+++ plom/scan/checkScanStatus.py
@@ -3,7 +3,7 @@
 # Copyright (C) 2020-2021 Colin B. Macdonald
 # Copyright (C) 2021 Jed Yeo
 
-import getpass
+from stdiomask import getpass
 
 from plom.misc_utils import format_int_list_with_runs
 from plom.messenger import ScanMessenger
@@ -20,7 +20,7 @@ def get_number_of_questions(server=None, pwd=None):
     msgr.start()
 
     if not pwd:
-        pwd = getpass.getpass("Please enter the 'scanner' password:")
+        pwd = getpass("Please enter the 'scanner' password: ")
 
     try:
         msgr.requestAndSaveToken("scanner", pwd)
@@ -32,15 +32,24 @@ def get_number_of_questions(server=None, pwd=None):
             "    e.g., on another computer?\n\n"
             'In order to force-logout the existing authorisation run "plom-scan clear"'
         )
-        exit(-1)
+        raise
 
-    spec = msgr.get_spec()
-    msgr.closeUser()
-    msgr.stop()
+    try:
+        spec = msgr.get_spec()
+    finally:
+        msgr.closeUser()
+        msgr.stop()
     return spec["numberOfQuestions"]
 
 
-def checkStatus(server=None, pwd=None):
+def check_and_print_scan_status(server=None, pwd=None):
+    """Prints summary of test/hw uploads.
+
+    More precisely. Prints lists
+    * which tests have been used (ie at least one uploaded page)
+    * which tests completely scanned (both tpages and hwpage)
+    * incomplete tests (missing one tpage or one hw-question)
+    """
     if server and ":" in server:
         s, p = server.split(":")
         msgr = ScanMessenger(s, port=p)
@@ -49,9 +58,8 @@ def checkStatus(server=None, pwd=None):
     msgr.start()
 
     if not pwd:
-        pwd = getpass.getpass("Please enter the 'scanner' password:")
+        pwd = getpass("Please enter the 'scanner' password: ")
 
-    # get started
     try:
         msgr.requestAndSaveToken("scanner", pwd)
     except PlomExistingLoginException as e:
@@ -62,15 +70,17 @@ def checkStatus(server=None, pwd=None):
             "    e.g., on another computer?\n\n"
             'In order to force-logout the existing authorisation run "plom-scan clear"'
         )
-        exit(10)
+        raise
 
-    spec = msgr.get_spec()
-
-    ST = msgr.getScannedTests()  # returns pairs of [page,version] - only display pages
-    UT = msgr.getUnusedTests()
-    IT = msgr.getIncompleteTests()
-    msgr.closeUser()
-    msgr.stop()
+    try:
+        spec = msgr.get_spec()
+        # returns pairs of [page,version] - only display pages
+        ST = msgr.getScannedTests()
+        UT = msgr.getUnusedTests()
+        IT = msgr.getIncompleteTests()
+    finally:
+        msgr.closeUser()
+        msgr.stop()
 
     print("Test papers unused: [{}]".format(format_int_list_with_runs(UT)))
 
@@ -126,7 +136,7 @@ def checkMissingHWQ(server=None, pwd=None):
     msgr.start()
 
     if not pwd:
-        pwd = getpass.getpass("Please enter the 'scanner' password:")
+        pwd = getpass("Please enter the 'scanner' password: ")
 
     try:
         msgr.requestAndSaveToken("scanner", pwd)
@@ -138,12 +148,13 @@ def checkMissingHWQ(server=None, pwd=None):
             "    e.g., on another computer?\n\n"
             'In order to force-logout the existing authorisation run "plom-scan clear"'
         )
-        exit(10)
+        raise
 
-    missingHWQ = msgr.getMissingHW()
-    msgr.closeUser()
-    msgr.stop()
-
+    try:
+        missingHWQ = msgr.getMissingHW()
+    finally:
+        msgr.closeUser()
+        msgr.stop()
     return missingHWQ
 
 
@@ -156,7 +167,7 @@ def replaceMissingHWQ(server, pwd, student_id, question):
     msgr.start()
 
     if not pwd:
-        pwd = getpass.getpass("Please enter the 'scanner' password:")
+        pwd = getpass("Please enter the 'scanner' password: ")
 
     try:
         msgr.requestAndSaveToken("scanner", pwd)
@@ -168,14 +179,14 @@ def replaceMissingHWQ(server, pwd, student_id, question):
             "    e.g., on another computer?\n\n"
             'In order to force-logout the existing authorisation run "plom-scan clear"'
         )
-        exit(10)
+        raise
 
-    rval = msgr.replaceMissingHWQuestion(
-        student_id=student_id, test=None, question=question
-    )  # can replace by SID or by test-number
-    msgr.triggerUpdateAfterHWUpload()
-
-    msgr.closeUser()
-    msgr.stop()
-
+    try:
+        rval = msgr.replaceMissingHWQuestion(
+            student_id=student_id, test=None, question=question
+        )  # can replace by SID or by test-number
+        msgr.triggerUpdateAfterHWUpload()
+    finally:
+        msgr.closeUser()
+        msgr.stop()
     return rval