+++ CHANGELOG.md
@@ -12,10 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `plom-server` has new command line args to control logging.
 * new `PlomServer`, `PlomDemoServer`, and `PlomLiteDemoServer` objects for interactively running servers, or otherwise running a server in a background process.
 * `plom-hwscan` can now specify precise per-page mappings from the bundle to questions on the server.
+* LaTeX errors are now displayed to markers.
 
 ### Changed
 * `plom-server launch --mastertoken aabbccdd...` replaces the old way (without the keyword argument).
 * QR creation now uses `segno` instead of `pyqrcode`.
+* `plom-hwscan` is more flexible about filenames: you do not need to put PDF files in a special directory.
+* The `plomgrading/server` container (Docker image) is now based on Ubuntu 20.04.
 
 ### Fixed
 * `plom-finish reassemble` not longer needs direct file access to the server (except when using the `--ided-only` option).